requirements.txt
pyyaml
jsonschema==4.19.0
jinja2
